refactor: split execute request hook and fade overlay out of the interactive editor
MultiCodeEditorInteractive crossed the 300-line react-doctor gate; the
parseCodeAndExecute callback moves body-intact into
interactive/use-execute-request.js (same bindings, same dependency
array, loading still reset in finally) and the FadeOverlay styled block
into interactive/fade-overlay.js, following the existing interactive/
decomposition pattern. The parent drops to 284 lines and no longer
imports @microlink/mql directly.
